MIMUW, Poland forecast
This is a late pull request because I had a bug in a script converting from ger-pol hub into the European one.
I am aware that the deadline is way behind us, but if it can be merged to make a full record I'd be grateful. 
I ensure that the predictions are based on the data up to the last Sunday.