Predictions from the last week from MIMUW
Guys, 
I am very sorry for the delay in the submission that comes from time to time - it is mostly due to the time lag in which I wait for the daily number of tests performed that we used in the model. I do my best not to make delays, but since the model evaluates few hours sometimes I collect the output Tuesday morning. 
I hope it will get merged, data used - only up to Sunday.