Add files via upload
The model performance may be affected due to the small number of hospitalized people with the omicron variant. That dramatically changed the ratio of hospitalized people on covid (due to severity of the disease) and with covid (due to other reasons), which affects the ascertainment rate estimate used in the model.