Add files via upload
The model performance may be affected due to the small number of hospitalized people with the omicron variant. That dramatically changed the ratio of hospitalized people on covid (due to severity of the disease) and with covid (due to other reasons), which affects the ascertainment rate estimate used in the model. Due to this unprecise estimate of ascertainment rate estimate and since the forecast created since 14-02-2022 proven imprecise, we are using a forecast created on 07-02-2022. We also identified a shift in the reported number of deaths (two weeks into the future) with the omicron variant that we are now using. We assume an increase in transmissibility and growth rate of a cluster in the second half of February due to the release of NPIs.