fix hospitalisation download
Hospitalisations data that are "stale" for a given location are excluded.

Up till now "stale" was >2 weeks old.

This broke down over the new year (with no new data since 18 Dec-5 Jan), producing a 0-row hospitalisation dataset. This in turn broke all the downstream uses of the data.

This PR updates this criteria to 1 month to solve this for now.

Downstream data uses are still quite vulnerable (I've fixed one to handle this error, but not others) but hopefully we are refactoring the hospitalisation download code soon anyway.