Update compile-model-reports.r
Changes the file to run model reports so that reports default to the latest available evaluation, instead of the current week evaluation.

At the moment if there isn't a current evaluation csv, the model report building action fails.

This failure is redundant because we should see any failure to produce the weekly evaluation csv in the "score" action.